refactor(Car): extract interface
- 테스트용 차량을 만들기 위해 move 메서드를 인터페이스로 추출
