+++ src/main/java/com/nextstep/racingcar/domain/round/Round.java
@@ -1,6 +1,6 @@
 package com.nextstep.racingcar.domain.round;
 
-import com.nextstep.racingcar.domain.car.Car;
+import com.nextstep.racingcar.domain.car.RandomMoveCar;
 import com.nextstep.racingcar.domain.car.Power;
 
 import java.util.ArrayList;
@@ -8,11 +8,11 @@ import java.util.List;
 import java.util.stream.IntStream;
 
 public class Round {
-    private List<Car> cars = new ArrayList<>();
+    private List<RandomMoveCar> randomMoveCars = new ArrayList<>();
 
     private Round(int numberOfCar) {
         IntStream.range(0, numberOfCar).forEach(number -> {
-            cars.add(new Car());
+            randomMoveCars.add(new RandomMoveCar());
         });
     }
 
@@ -21,10 +21,10 @@ public class Round {
     }
 
     public int getCarNumber() {
-        return this.cars.size();
+        return this.randomMoveCars.size();
     }
 
     public void moveAll() {
-        this.cars.forEach(car -> car.move(Power.generate()));
+        this.randomMoveCars.forEach(randomMoveCar -> randomMoveCar.move(Power.generate()));
     }
 }
